requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-yasg==1.21.7
django-filter==24.3
django-environ==0.11.2
Pillow==10.4.0
